requirements.txt
altair==4.2.2
streamlit==1.15.0
Pillow==9.2.0
langchain==0.0.229
langchain-google-genai==0.0.1
pytesseract==0.3.10  # => is python wrapper tool
gTTS==2.3.0
google-api-python-client==2.79.0
google-auth==2.18.0
requests==2.28.1
numpy==1.24.2
pip==24.3.1
setuptools==65.5.0
torch==2.0.1
torchvision==0.15.2